[svn-r2148] Corrected a few problems in the free-list code and added more assert() macros to double-check things.  I've turned them back on again now.  I also changed the internal representation of a few struct fields to be float instead of double, since the HP/UX 10.20 compiler was having problems with the alignment of the doubles.

+++ src/H5Bprivate.h
@@ -132,7 +132,7 @@ __DLL__ herr_t H5B_create (H5F_t *f, const H5B_class_t *type, void *udata,
 __DLL__ herr_t H5B_find (H5F_t *f, const H5B_class_t *type,
 			 const haddr_t *addr, void *udata);
 __DLL__ herr_t H5B_insert (H5F_t *f, const H5B_class_t *type,
-			   const haddr_t *addr, const double split_ratios[],
+			   const haddr_t *addr, const float split_ratios[],
 			   void *udata);
 __DLL__ herr_t H5B_remove(H5F_t *f, const H5B_class_t *type,
 			  const haddr_t *addr, void *udata);
